requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
urllib3>=2.0.0
python-dotenv==1.2.1
